+++ controllers/AboutController.js
@@ -1,0 +1,8 @@
+export default class AboutController {
+    constructor (){
+        this.load();
+    }
+    load(){
+        
+    }
+}
