+++ server/models/post.js
@@ -1,12 +1,11 @@
 const mongoose = require("mongoose");
 
- const postSchema = new mongoose.Schema({
-   userId: String,
-   imagePath: String,
-   title: String,
-   description: String,
-   date: Date
+const postSchema = new mongoose.Schema({
+    userId: String,
+    imagePath: String,
+    title: String,
+    description: String,
+    date: Date
+})
 
- })
-
- module.exports = mongoose.model("Post", postSchema)
+module.exports = mongoose.model("Post", postSchema);
