+++ templates/add-post.html
@@ -1,5 +1,5 @@
 <div>
-    <form id="sendPost" encrtype="multipart/form-data">
+    <form id="sendPost" enctype="multipart/form-data">
         <div>
             <div>
                 <span>Title: </span> <input class="form-input-card" type="text" name="title">
